Revert "Strip out unused symbols for the bfloat utilties (#34364)"
This reverts commit a4347c0dda43c08c3953ac254e90b2122ed45901.
