+++ tests/tt_metal/tt_metal/api/test_blockfloat_common.cpp
@@ -4,7 +4,7 @@
 
 #include <gtest/gtest.h>
 #include <cstdint>
-#include "impl/data_format/blockfloat_common.hpp"
+#include <tt-metalium/blockfloat_common.hpp>
 #include <bit>
 #include <memory>
