+++ tests/tt_metal/tt_metal/integration/test_flatten.cpp
@@ -30,7 +30,6 @@
 #include <tt-metalium/program.hpp>
 #include <tt_stl/span.hpp>
 #include <tt-metalium/tt_backend_api_types.hpp>
-#include "impl/data_format/bfloat16_utils.hpp"
 
 namespace tt::tt_metal {
