+++ tests/tt_metal/tt_metal/perf_microbenchmark/old/noc/test_noc_read_global_l1.cpp
@@ -40,7 +40,6 @@
 #include "tt_metal/test_utils/deprecated/tensor.hpp"
 #include <tt-metalium/mesh_device.hpp>
 #include <tt-metalium/distributed.hpp>
-#include "impl/data_format/bfloat16_utils.hpp"
 
 using namespace tt;
 using std::chrono::duration_cast;
