+++ tests/tt_metal/tt_metal/perf_microbenchmark/old/noc/test_noc_read_local_l1.cpp
@@ -37,7 +37,6 @@
 #include "tt_metal/test_utils/deprecated/tensor.hpp"
 #include <tt-metalium/mesh_device.hpp>
 #include <tt-metalium/distributed.hpp>
-#include "impl/data_format/bfloat16_utils.hpp"
 
 using namespace tt;
 using std::chrono::duration_cast;
