+++ tt_metal/api/tt-metalium/bfloat16.hpp
@@ -65,11 +65,19 @@ bool operator==(const std::vector<bfloat16>& lhs, const std::vector<bfloat16>& r
 
 uint32_t pack_two_bfloat16_into_uint32(std::pair<bfloat16, bfloat16> two_bfloats);
 
+std::pair<bfloat16, bfloat16> unpack_two_bfloat16_from_uint32(uint32_t uint32_data);
+
+std::vector<std::uint32_t> create_arange_vector_of_bfloat16(size_t num_bytes, bool print = true);
+
 std::vector<bfloat16> create_random_vector_of_bfloat16_native(
     size_t num_bytes, float rand_max_float, int seed, float offset = 0.0f);
 
 std::vector<std::uint32_t> create_random_vector_of_bfloat16(
     size_t num_bytes, int rand_max_float, int seed, float offset = 0.0f);
+
+std::vector<std::uint32_t> create_random_vector_of_bfloat16_1_1(size_t num_bytes, int seed);
+
+std::vector<std::uint32_t> create_random_vector_of_bfloat16_0_2(size_t num_bytes, int seed);
 
 /*
  * rk: Still won't handle the case where the number of elements is odd, except
@@ -80,6 +88,22 @@ std::vector<std::uint32_t> create_constant_vector_of_bfloat16(size_t num_bytes, 
 // creates a bfloat16 identity matrix with dims (rows x cols)
 // each 2 cols will be packed as a single uint32_t
 std::vector<bfloat16> create_identity_matrix(int rows, int cols, int num_ones);
+
+// TODO(AP): duplication with above
+std::vector<uint32_t> create_random_binary_vector_of_bfloat16(size_t num_bytes, int seed);
+
+std::vector<uint16_t> u16_from_u32_vector(const std::vector<uint32_t>& in);
+
+std::vector<uint32_t> u32_from_u16_vector(const std::vector<uint16_t>& in);
+
+void print_vec_of_uint32_as_packed_bfloat16(
+    const std::vector<std::uint32_t>& vec, int num_tiles, const std::string& name = "", int tile_print_offset = 0);
+
+void print_vec_of_bfloat16(
+    const std::vector<bfloat16>& vec, int num_tiles, const std::string& name = "", int tile_print_offset = 0);
+
+void print_vec(
+    const std::vector<uint32_t>& vec, int num_tiles, const std::string& name = "", int tile_print_offset = 0);
 
 std::vector<uint32_t> pack_bfloat16_vec_into_uint32_vec(const std::vector<bfloat16>& data);
 
@@ -92,5 +116,13 @@ std::vector<bfloat16> unpack_uint32_vec_into_bfloat16_vec(
 // Equality functions
 bool equal_within_n_sig_figs(float a, float b, int n);
 
+bool equal_within_absolute_tolerance(float a, float b, float tol);
+
 // this follows the implementation of numpy's is_close
 bool is_close(float a, float b, float rtol = 0.01f, float atol = 0.001f);
+
+bool packed_uint32_t_vector_comparison(
+    const std::vector<uint32_t>& vec_a,
+    const std::vector<uint32_t>& vec_b,
+    const std::function<bool(float, float)>& comparison_function,
+    int* argfail = nullptr);
