+++ tt_metal/api/tt-metalium/bfloat4.hpp
@@ -11,6 +11,21 @@
 #include <optional>
 #include <vector>
 
+constexpr int log2(int n) {
+    int log = 0;
+    while (n >>= 1) {
+        ++log;
+    }
+    return log;
+}
+
+[[deprecated("Use pack_as_bfp4_tiles instead.")]]
+std::vector<uint32_t> pack_fp32_vec_as_bfp4_tiles(
+    tt::stl::Span<const float> fp32_vec,
+    bool row_major_input,
+    bool is_exp_a,
+    const std::optional<tt::tt_metal::Tile>& tile = std::nullopt);
+
 template <typename T>
 std::vector<uint32_t> pack_as_bfp4_tiles(
     tt::stl::Span<const T> data,
