+++ tt_metal/impl/data_format/bfloat4.cpp
@@ -21,7 +21,13 @@
 #include "tracy/Tracy.hpp"
 #include "tt_backend_api_types.hpp"
 
-constexpr int log2(int n) { return (n <= 1) ? 0 : std::bit_width(static_cast<unsigned>(n)) - 1; }
+std::vector<uint32_t> pack_fp32_vec_as_bfp4_tiles(
+    tt::stl::Span<const float> fp32_vec,
+    bool row_major_input,
+    bool is_exp_a,
+    const std::optional<tt::tt_metal::Tile>& tile) {
+    return pack_as_bfp_tiles<tt::DataFormat::Bfp4_b>(fp32_vec, row_major_input, is_exp_a, tile);
+}
 
 template <typename T>
 std::vector<uint32_t> pack_as_bfp4_tiles(
