+++ tt_metal/impl/debug/dprint_server.cpp
@@ -26,7 +26,7 @@
 
 #include <enchantum/enchantum.hpp>
 #include <tt-logger/tt-logger.hpp>
-#include "impl/data_format/blockfloat_common.hpp"
+#include <tt-metalium/blockfloat_common.hpp>
 #include <tt_stl/assert.hpp>
 #include <umd/device/types/core_coordinates.hpp>
 #include <umd/device/soc_descriptor.hpp>
